Add files via upload
It was one of the best work that i think I had done in my education time. It was an application which takes of test of patient and predictive the possibility of patient, will attend the icu in near future or not. I got dataset from kaggle ( uploaded by Mexican govt. ), I clean the dataset and fed to ML Algorithms such as decision forest etc. this model does not perform well on accuracy metrix but perform up-to mark on z1 score.